% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvision_init.R
\name{gvision_init}
\alias{gvision_init}
\title{authorization for Google Vision}
\usage{
gvision_init()
}
\value{
nothing.
}
\description{
Initializes the authorization credentials for the 'Google Vision' API.
Needs to be called before using any other functions of \code{imgrec} and requires \code{gvision_key} as environment variable.
}
\examples{
\dontrun{
Sys.setenv(gvision_key = "Your Google Vision API key")

gvision_init()
}




}
